pymemo/requirements.txt
datetime
wheel
imutils
numpy
opencv-python
setuptools
twine 
pyyaml
